+++ framework/db-local/src/main/java/is/codion/framework/db/local/DefaultLocalEntityConnection.java
@@ -171,7 +171,7 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection {
 	}
 
 	@Override
-	public void beginTransaction() {
+	public void startTransaction() {
 		synchronized (connection) {
 			connection.startTransaction();
 		}