+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntitySearchModelTest.java
@@ -251,7 +251,7 @@ public final class DefaultEntitySearchModelTest {
 						.columns(searchable)
 						.build();
 
-		CONNECTION_PROVIDER.connection().beginTransaction();
+		CONNECTION_PROVIDER.connection().startTransaction();
 		setupData();
 	}